Moved Encoded and Decoder to root of component directory
